move api declarations to separate file
